split version cont for build
