+++ infrastructure/src/shared/const.ts
@@ -1,3 +1,2 @@
 export const KEY_ALIAS = 'sd-extension-password-key';
 export const STACK_ID = 'Extension-for-Stable-Diffusion-on-AWS';
-export const ECR_VERSION: string = 'v1.5.0-ca5c221';
