+++ infrastructure/src/shared/version.ts
@@ -1,0 +1,1 @@
+export const ECR_VERSION: string = 'v1.5.0-ca5c221';
